+++ src/Faultline.Web/Program.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 using Faultline.Web;
 using Faultline.Web.Shell;
+using Faultline.Web.Shell.Playtest;
 
 var builder = WebAssemblyHostBuilder.CreateDefault(args);
 builder.RootComponents.Add<App>("#app");
@@ -14,5 +15,6 @@ builder.Services.AddSingleton<CustomFightStore>();
 builder.Services.AddSingleton<PlaytestNotes>();
 builder.Services.AddSingleton<RunStore>();
 builder.Services.AddSingleton<RunSession>();
+builder.Services.AddSingleton<BoardAnimator>();
 
 await builder.Build().RunAsync();